Refactored IF domain objects into logical order of case classes as this is important!
