--- app/uk/gov/hmrc/individualsincomeapi/domain/integrationframework/paye/IFEmployeePensionContribs.scala
@@ -1,47 +1,0 @@
-/*
- * Copyright 2020 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package uk.gov.hmrc.individualsincomeapi.domain.integrationframework.paye
-
-import play.api.libs.functional.syntax.{unlift, _}
-import play.api.libs.json.{Format, JsPath}
-import uk.gov.hmrc.individualsincomeapi.domain.integrationframework.paye.IFIncomePaye._
-
-case class IFEmployeePensionContribs(
-  paidYtd: Option[Double],
-  notPaidYtd: Option[Double],
-  paid: Option[Double],
-  notPaid: Option[Double]
-)
-
-object IFEmployeePensionContribs {
-
-  implicit val employeePensionContribsFormat: Format[IFEmployeePensionContribs] = Format(
-    (
-      (JsPath \ "paidYTD").readNullable[Double](paymentAmountValidator) and
-        (JsPath \ "notPaidYTD").readNullable[Double](paymentAmountValidator) and
-        (JsPath \ "paid").readNullable[Double](paymentAmountValidator) and
-        (JsPath \ "notPaid").readNullable[Double](paymentAmountValidator)
-    )(IFEmployeePensionContribs.apply _),
-    (
-      (JsPath \ "paidYTD").writeNullable[Double] and
-        (JsPath \ "notPaidYTD").writeNullable[Double] and
-        (JsPath \ "paid").writeNullable[Double] and
-        (JsPath \ "notPaid").writeNullable[Double]
-    )(unlift(IFEmployeePensionContribs.unapply))
-  )
-
-}
