--- app/uk/gov/hmrc/individualsincomeapi/domain/integrationframework/paye/IFGrossEarningsForNics.scala
@@ -1,47 +1,0 @@
-/*
- * Copyright 2020 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package uk.gov.hmrc.individualsincomeapi.domain.integrationframework.paye
-
-import play.api.libs.json.{Format, JsPath}
-import uk.gov.hmrc.individualsincomeapi.domain.integrationframework.paye.IFIncomePaye._
-import play.api.libs.functional.syntax.{unlift, _}
-
-case class IFGrossEarningsForNics(
-  inPayPeriod1: Option[Double],
-  inPayPeriod2: Option[Double],
-  inPayPeriod3: Option[Double],
-  inPayPeriod4: Option[Double]
-)
-
-object IFGrossEarningsForNics {
-
-  implicit val grossEarningsForNicsFormat: Format[IFGrossEarningsForNics] = Format(
-    (
-      (JsPath \ "inPayPeriod1").readNullable[Double](paymentAmountValidator) and
-        (JsPath \ "inPayPeriod2").readNullable[Double](paymentAmountValidator) and
-        (JsPath \ "inPayPeriod3").readNullable[Double](paymentAmountValidator) and
-        (JsPath \ "inPayPeriod4").readNullable[Double](paymentAmountValidator)
-    )(IFGrossEarningsForNics.apply _),
-    (
-      (JsPath \ "inPayPeriod1").writeNullable[Double] and
-        (JsPath \ "inPayPeriod2").writeNullable[Double] and
-        (JsPath \ "inPayPeriod3").writeNullable[Double] and
-        (JsPath \ "inPayPeriod4").writeNullable[Double]
-    )(unlift(IFGrossEarningsForNics.unapply))
-  )
-
-}
