+++ test/unit/uk/gov/hmrc/individualsincomeapi/domain/integrationframework/paye/IFEmployeePensionContribsSpec.scala
@@ -16,11 +16,12 @@
 
 package unit.uk.gov.hmrc.individualsincomeapi.domain.integrationframework.paye
 
+import org.scalatest.{Matchers, WordSpec}
 import play.api.libs.json.Json
 import uk.gov.hmrc.individualsincomeapi.domain.integrationframework.paye.IFEmployeePensionContribs
-import utils.SpecBase
+import uk.gov.hmrc.individualsincomeapi.domain.integrationframework.paye.IFPayeEntry._
 
-class IFEmployeePensionContribsSpec extends SpecBase {
+class IFEmployeePensionContribsSpec extends WordSpec with Matchers {
 
   val validEmployeePensionContribs = IFEmployeePensionContribs(
     Some(169731.51),
