+++ test/utils/IncomePayeHelpers.scala
@@ -16,8 +16,7 @@
 
 package utils
 
-import uk.gov.hmrc.individualsincomeapi.domain._
-import uk.gov.hmrc.individualsincomeapi.domain.integrationframework.paye.{IFBenefits, IFEmployeeNics, IFEmployeePensionContribs, IFPayeEntry, IFPostGradLoan, IFStudentLoan}
+import uk.gov.hmrc.individualsincomeapi.domain.integrationframework.paye.{IFBenefits, IFEmployeeNics, IFEmployeePensionContribs, IFGrossEarningsForNics, IFPayeEntry, IFPostGradLoan, IFStudentLoan, IFTotalEmployerNics}
 
 trait IncomePayeHelpers {
   def createValidPayeEntry() =
@@ -27,6 +26,7 @@ trait IncomePayeHelpers {
       Some(19157.5),
       Some(3095.89),
       Some(159228.49),
+      Some(createValodIFGrossEarningsForNics),
       Some("345/34678"),
       Some("2006-02-27"),
       Some(16533.95),
@@ -35,6 +35,7 @@ trait IncomePayeHelpers {
       Some("2"),
       Some("W4"),
       Some(198035.8),
+      Some(createValidTotalEmployerNics()),
       Some(createValidEmployeeNics()),
       Some(createValidEmployeePensionContribs()),
       Some(createValidBenefits()),
@@ -55,6 +56,18 @@ trait IncomePayeHelpers {
       Some(162081.23)
     )
 
+  private def createValidTotalEmployerNics() =
+    IFTotalEmployerNics(
+      Some(15797.45),
+      Some(13170.69),
+      Some(16193.76),
+      Some(30846.56),
+      Some(10633.5),
+      Some(15579.18),
+      Some(110849.27),
+      Some(162081.23)
+    )
+
   private def createValidEmployeePensionContribs() =
     IFEmployeePensionContribs(Some(169731.51), Some(173987.07), Some(822317.49), Some(818841.65))
 
@@ -63,4 +76,7 @@ trait IncomePayeHelpers {
   private def createValidStudentLoan() = IFStudentLoan(Some("02"), Some(88478.16), Some(545.52))
 
   private def createValidPostGradLoan() = IFPostGradLoan(Some(15636.22), Some(46849.26))
+
+  private def createValodIFGrossEarningsForNics() =
+    IFGrossEarningsForNics(Some(169731.51), Some(173987.07), Some(822317.49), Some(818841.65))
 }
